@@ -17,24 +17,42 @@
     "flux": "^3.1.3",
     "method-override": "^2.3.10",
     "pug": "^2.0.0-rc.4",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-ga": "^2.3.5",
     "react-router": "^3.2.0",
-    "webpack": "^1.15.0"
+    "webpack": "^4.35.0"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^6.4.1",
-    "babel-preset-es2015": "^6.24.1",
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
+    "@babel/preset-react": "^7.0.0",
+    "autoprefixer": "^9.6.0",
+    "babel-loader": "^8.0.6",
     "babel-preset-react": "^6.24.1",
-    "grunt": "^1.0.1",
-    "grunt-contrib-clean": "^1.1.0",
-    "grunt-contrib-cssmin": "^1.0.2",
-    "grunt-contrib-uglify": "^2.0.0",
-    "grunt-modernizr": "^1.0.2",
-    "grunt-nodemon": "^0.4.2",
-    "grunt-webpack": "^1.0.18",
-    "webpack-dev-server": "^1.16.5"
+    "browser-sync": "^2.26.7",
+    "browsersync": "^0.0.1-security",
+    "child_process": "^1.0.2",
+    "cp": "^0.2.0",
+    "cssnano": "^4.1.10",
+    "del": "^4.1.1",
+    "eslint": "^5.16.0",
+    "gulp": "^4.0.2",
+    "gulp-cssmin": "^0.2.0",
+    "gulp-eslint": "^5.0.0",
+    "gulp-imagemin": "^6.0.0",
+    "gulp-newer": "^1.4.0",
+    "gulp-plumber": "^1.2.1",
+    "gulp-postcss": "^8.0.0",
+    "gulp-rename": "^1.4.0",
+    "gulp-sass": "^4.0.2",
+    "imagemin": "^6.1.0",
+    "newer": "^0.1.5",
+    "plumber": "^0.4.8",
+    "postcss": "^7.0.17",
+    "rename": "^1.0.4",
+    "sass": "^1.21.0",
+    "webpack-dev-server": "^1.16.5",
+    "webpack-stream": "^5.2.1"
   }
 }
